+++ agape/src/widgets/image.rs
@@ -80,7 +80,7 @@ impl Widget for Image {
         self.id
     }
 
-    fn layout(&self) -> Box<dyn Layout> {
+    fn layout(&self, _: &mut Renderer) -> Box<dyn Layout> {
         let layout = EmptyLayout {
             id: self.id,
             intrinsic_size: self.style.intrinsic_size.clone(),